requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.34.0
scikit-learn==1.4.2
pandas==2.2.2
python-dotenv==1.0.1
joblib==1.4.2
plotly==5.22.0
requests==2.31.0
sqlalchemy==2.0.30
google-generativeai==0.7.2
pydantic==2.7.1